% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{applyFormula}
\alias{applyFormula}
\title{Apply formulas}
\usage{
applyFormula(mod, env)
}
\arguments{
\item{mod}{Model}

\item{env}{Environment, typically a named list.}
}
\value{
A vector of size nFormula, containing the result of evaluated formulas
}
\description{
Attempt to compute all formulas in a model, using parameter/process/covariate
values contained in an environment env.
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=rlnorm(50),X=data.frame(x1=rnorm(50),x2=runif(50)))
# Define parameters to be estimated: mu0, mu1 , mu2 and sigma
param <- list(parameter(name='mu0',init=0),parameter(name='mu1',init=0),
              parameter(name='mu2',init=0),parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu0+mu1*x1+mu2*x2','sdlog=sigma'),par=param)
# Evaluate formula
applyFormula(mod,env=list(mu0=0,mu1=0.1,mu2=-0.1,x1=1,x2=0.4,sigma=0.5))
# If environment is incomplete (here: x1 missing), a NA is returned for unevaluated formula
applyFormula(mod,env=list(mu0=0,mu1=0.1,mu2=-0.1,x2=0.4,sigma=0.5))
}
