% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dimension.R
\name{as.data.frame.dimension}
\alias{as.data.frame.dimension}
\title{Dimension to data frame}
\usage{
\method{as.data.frame}{dimension}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Dimension object, dataset to be converted.}

\item{row.names}{NULL or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical, see ?as.data.frame. Unused here.}

\item{...}{Optional arguments.}
}
\value{
A data frame merging coordinates and covariates of the dimension.
}
\description{
Convert an object of class 'dimension' into a data frame
}
\examples{
as.data.frame(dimension(name='time',coord=data.frame(t=1:50)))
}
