% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_distance.R
\name{distance}
\alias{distance}
\title{Distance constructor.}
\usage{
distance(funk = "Euclidean", par = NULL)
}
\arguments{
\item{funk}{Character, distance function.
getCatalogue(printOnly=TRUE) for a list of available distance functions}

\item{par}{Parameter, a list of parameter objects.}
}
\value{
An object of class 'distance'.
}
\description{
Creates a new instance of a 'distance' object
}
\examples{
d1 <- distance(funk='Euclidean')
d2 <- distance(funk='Haversine')
}
