% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elNino}
\alias{elNino}
\title{Nino 3.4 spring index, 1951-2014}
\format{
A data frame with 64 rows and 2 variables:
\describe{
  \item{year}{year from 1951 to 2014}
  \item{nino}{nino 3.4 index}
}
}
\source{
\url{https://www.esrl.noaa.gov/psd/gcos_wgsp/Timeseries/Data/nino34.long.data}
}
\usage{
elNino
}
\description{
A dataset containing standardized sea surface temperature anomalies
during the spring (September - October - November)
over the Nino 3.4 region for the period 1951-2014.
See \url{https://climatedataguide.ucar.edu/climate-data/nino-sst-indices-nino-12-3-34-4-oni-and-tni}
}
\keyword{datasets}
