% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getCatalogue}
\alias{getCatalogue}
\title{STooDs catalogue}
\usage{
getCatalogue(printOnly = FALSE)
}
\arguments{
\item{printOnly}{Logical, should the catalogue be returned or only printed?}
}
\value{
If \code{printOnly==FALSE}, a list with the following fields:
\describe{
  \item{distribution}{available univariate distributions.}
  \item{mvtdistribution}{available multivariate distributions.}
  \item{distance}{available distances.}
}
}
\description{
Distributions, multivariate distributions and distances available in STooDs
}
\examples{
catalogue <- getCatalogue()
getCatalogue(printOnly=TRUE)
}
