% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{getDAG}
\alias{getDAG}
\title{Model DAG}
\usage{
getDAG(
  mod,
  expandData = TRUE,
  expandProcess = TRUE,
  nmax = 100,
  dataNodeSize = 5,
  dparNodeSize = 12,
  parNodeSize = dparNodeSize,
  processNodeSize = dataNodeSize,
  covNodeSize = dparNodeSize,
  dimNodeSize = 25,
  minNodeSize = 3
)
}
\arguments{
\item{mod}{Model}

\item{expandData}{logical. If true each data point defines a node}

\item{expandProcess}{logical. If true each process value defines a node}

\item{nmax}{integer. When expandData and/or expandProcess is true, nmax
defines the maximum number of nodes for individual data
or process values}

\item{dataNodeSize}{numerix. Size of a data node.}

\item{dparNodeSize}{logical. Size of parent distribution parameter node.}

\item{parNodeSize}{logical. Size of a parameter node.}

\item{processNodeSize}{logical. Size of a process node.}

\item{covNodeSize}{logical. Size of a covariate node.}

\item{dimNodeSize}{logical. Size of a dimension node.}

\item{minNodeSize}{logical. Minimum node size. Used for dataset/process
node when expandData/expandProcess is TRUE.}
}
\value{
A DAG, i.e. a list with 2 data frames: nodes and links.
}
\description{
Get the Directed Acyclic Graph  of a model
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=data.frame(Y=rlnorm(50)))
# Define parameters to be estimated: mu and sigma
param <- list(parameter(name='mu',init=0),parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu','sdlog=sigma'),par=param)
# get its DAG
dag<-getDAG(mod)
}
