% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getNames.default}
\alias{getNames.default}
\title{Default getNames from an object or a list of objects having a $name field
(e.g. parameters,processes)}
\usage{
\method{getNames}{default}(loo)
}
\arguments{
\item{loo}{List Of Objects}
}
\value{
A character vector containg names
}
\description{
Default getNames from an object or a list of objects having a $name field
(e.g. parameters,processes)
}
