% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getNames}
\alias{getNames}
\title{Generic getNames function}
\usage{
getNames(o, ...)
}
\arguments{
\item{o}{Object}

\item{...}{Optional arguments.}
}
\value{
A character vector containg names
}
\description{
Generic getNames function
}
