% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getParentDist}
\alias{getParentDist}
\title{Get parent distribution associated with row(s) in the dataset}
\usage{
getParentDist(model, estimate, rows = 1:NROW(model$dataset$Y))
}
\arguments{
\item{model}{Model}

\item{estimate}{Named vector}

\item{rows}{Integer vector, rows in the dataset}
}
\value{
A list of size nRows; each element is itself a list with fields
   $var (variable name), $dist (character, distribution name)
   and $par (numeric vector, parameters)
}
\description{
Get parent distribution (name + parameters) associated with
rows in the dataset, using values contained in a named vector 'estimate'
(typically: posterior max or median).
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=rlnorm(50),X=1:50)
# Define parameters to be estimated: mu0, mu1 and sigma
param <- list(parameter(name='mu0',init=0),parameter(name='mu1',init=0),
              parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu0+mu1*X','sdlog=sigma'),par=param)
# Get environment
res <- getParentDist(mod,estimate=list(mu0=0,mu1=1,sigma=1))
i=1;paste0('Distribution: ',res[[i]]$dist,' - parameters: ',paste(res[[i]]$par,collapse=','))
i=10;paste0('Distribution: ',res[[i]]$dist,' - parameters: ',paste(res[[i]]$par,collapse=','))
i=50;paste0('Distribution: ',res[[i]]$dist,' - parameters: ',paste(res[[i]]$par,collapse=','))
}
