% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{getRowEnv}
\alias{getRowEnv}
\title{Get environment associated with a row in the dataset}
\usage{
getRowEnv(model, estimate, row)
}
\arguments{
\item{model}{Model}

\item{estimate}{Named vector}

\item{row}{Integer, row in the dataset}
}
\value{
A list containing the environment
}
\description{
Get environment (i.e. parameter/process/covariate values) associated with
a row in the dataset, using values contained in a named vector 'estimate'
(typically: posterior max or median).
For instance if 'estimate' contains columns named 'p_1',...'p_50' corresponding
to process values at 50 sites, this function retrieves the site associated with
'row' and extract the corresponding 'p' value.
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=rlnorm(50),X=1:50)
# Define parameters to be estimated: mu0, mu1 and sigma
param <- list(parameter(name='mu0',init=0),parameter(name='mu1',init=0),
              parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu0+mu1*X','sdlog=sigma'),par=param)
# Get environment
getRowEnv(mod,estimate=list(mu0=0,mu1=0.1,sigma=1),row=1)
getRowEnv(mod,estimate=list(mu0=0,mu1=0.1,sigma=1),row=10)
}
