% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dimension.R
\name{is.dimension}
\alias{is.dimension}
\title{Dimension tester.}
\usage{
is.dimension(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'dimension', FALSE otherwise.
}
\description{
Is an object of class 'dimension'?
}
\keyword{internal}
