% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_distance.R
\name{is.distance}
\alias{is.distance}
\title{Distance tester.}
\usage{
is.distance(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'distance', FALSE otherwise.
}
\description{
Is an object of class 'distance'?
}
\keyword{internal}
