% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmc.R
\name{is.mcmc}
\alias{is.mcmc}
\title{mcmc tester}
\usage{
is.mcmc(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'mcmc', FALSE otherwise.
}
\description{
Is an object of class 'mcmc'?
}
\keyword{internal}
