% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_process.R
\name{is.process}
\alias{is.process}
\title{Process tester.}
\usage{
is.process(o)
}
\arguments{
\item{o}{Object, an object.}
}
\value{
A logical equal to TRUE if class(o)== 'process', FALSE otherwise.
}
\description{
Is an object of class 'process'?
}
\keyword{internal}
