% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MCMC object constructor.}
\usage{
mcmc(
  Nsim = 10000,
  Nadapt = 50,
  stopAdapt = 5000,
  Nburn = 0,
  Nslim = 1,
  minMR = 0.1,
  maxMR = 0.5,
  downMult = 0.9,
  upMult = 1.1,
  fileName = "MCMC.txt",
  useSpeedUp = TRUE,
  jumpFileName = "",
  moveRateFileName = ""
)
}
\arguments{
\item{Nsim}{Integer, total number of iterations.}

\item{Nadapt}{Integer, adaptation period: jump sizes are increased/decreased
every Nadapt iterations to comply with the desired moving rates.}

\item{stopAdapt}{Integer, no more adaptation after iteration number stopAdapt.}

\item{Nburn}{Integer, number of inital iterations that will not be written to file.}

\item{Nslim}{Integer, only one iteration very Nslim will be written to file.}

\item{minMR}{Numeric in (0;1), lower bound for the desired move rate interval.}

\item{maxMR}{Numeric in (0;1), upper bound for the desired move rate interval.}

\item{downMult}{Numeric in (0:1), multiplication factor used to decrease jump size
when move rate is too low.}

\item{upMult}{Numeric (>1, avoid 1/dowMult) multiplication factor used to increase
jump size when move rate is too high.}

\item{fileName}{Character, name of MCMC file (|!| name of the file only, not full path).}

\item{useSpeedUp}{Logical, enable all computational tricks to optimize speed?
Should always be .true., except if one wishes to verify the equivalence speedUp/no speedUp.}

\item{jumpFileName}{Character, name of file reporting MCMC jump sizes at each adaptation
(|!| name of the file only, not full path). No reporting if empty.}

\item{moveRateFileName}{Character, name of file reporting MCMC move rates at each adaptation
(|!| name of the file only, not full path). No reporting if empty.}
}
\value{
An object of class 'mcmc'.
}
\description{
Creates a new instance of a 'mcmc' object
}
\examples{
m <- mcmc()
}
