% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{MCMCreport}
\alias{MCMCreport}
\title{MCMC reporting}
\usage{
MCMCreport(
  mcmc,
  model,
  traceFile = "MCMCtrace.pdf",
  parFile = "MCMCpar.pdf",
  processFile1 = "MCMCprocess.pdf",
  processFile2 = "MCMCprocess2.pdf",
  panelPerCol = 10,
  panelHeight = 3,
  panelWidth = 23/panelPerCol,
  addWorld = FALSE
)
}
\arguments{
\item{mcmc}{data frame, MCMC samples (typically from a call to readMCMC())}

\item{model}{Model object, the model that has been MCMC-sampled.}

\item{traceFile}{Character, full path to traceplots pdf file, not created if NULL}

\item{parFile}{Character, full path to parplots pdf file, not created if NULL}

\item{processFile1}{Character, full path to 1st processplot pdf file, not created if NULL}

\item{processFile2}{Character, full path to 2nd processplot pdf file, not created if NULL}

\item{panelPerCol}{Integer, max number of panels per column}

\item{panelHeight}{Numeric, height of each panel}

\item{panelWidth}{Numeric, width of each panel}

\item{addWorld}{Logical, add world contours for spatial processes?}
}
\description{
Generate pdf report files summarizing mcmc samples
}
