% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{model}
\alias{model}
\title{Model constructor.}
\usage{
model(
  dataset,
  parentDist,
  varName = ifelse(length(parentDist) == 1, as.character(dataset$var[1, 1]), NA),
  formula = NULL,
  par = NULL,
  process = NULL,
  name = "STooDs_model"
)
}
\arguments{
\item{dataset}{Dataset object.}

\item{parentDist}{Character vector, name of the parent distribution(s).
In cases where the predictand is multivariate, should be of size nVar.}

\item{varName}{Character vector, variable name(s), size nVar.
Compulsory for a multivariate predictand (nVar > 1)}

\item{formula}{Character vector, formula(s) to retrieve each parameter
of the parent distribution(s). Size sum(nParentPar).}

\item{par}{List of parameter objects, containing the parameters usable
in the formulas.}

\item{process}{List of Process objects, containing the processes usable
in the formulas.}

\item{name}{Character, name / description of the model}
}
\value{
An object of class 'model'.
}
\description{
Creates a new instance of a 'model' object
}
\examples{
# Example of a very simple model, see case studies for more varied examples.
# Generate a fake dataset
dat <- dataset(Y=data.frame(Y=rlnorm(50)))
# Define parameters to be estimated: mu and sigma
param <- list(parameter(name='mu',init=0),parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu','sdlog=sigma'),par=param)
}
