% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OZstations}
\alias{OZstations}
\title{42 Hydrometric stations in Eastern Australia}
\format{
A data frame with 42 rows and 5 variables:
\describe{
  \item{ID}{Identifier}
  \item{OzID}{Original identifier}
  \item{lon}{longitude (°)}
  \item{lat}{latitude (°)}
  \item{area}{catchment area (km2)}
}
}
\source{
\url{http://www.bom.gov.au/water/hrs/}
}
\usage{
OZstations
}
\description{
A dataset describing 42 hydrometric stations in Eastern Australia
}
\keyword{datasets}
