% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dimension.R
\name{plot.dimension}
\alias{plot.dimension}
\title{plot a Dimension}
\usage{
\method{plot}{dimension}(
  x,
  col = "red",
  alpha = 0.7,
  size = 3,
  type = "l",
  palette = "Spectral",
  colMax = 4,
  addWorld = FALSE,
  ...
)
}
\arguments{
\item{x}{Dimension object.}

\item{col}{Color. color of the points or lines}

\item{alpha}{Numeric. transparency}

\item{size}{Numeric. point size}

\item{type}{character. 'l' (line), 'p' (points) or 'b' (both)}

\item{palette}{character. Name of the color distiller palette for coloring points on a map}

\item{colMax}{Integer.  Max number of columns in plot matrix}

\item{addWorld}{Logical. Add world contours?}

\item{...}{Optional arguments.}
}
\value{
nothing - plotting function.
}
\description{
Plot a Dimension object (coordinates and covariates)
}
\examples{
# Define dimensions
time=dimension(name='time',coord=data.frame(t=1:50),z=data.frame(ft=rnorm(50)))
space=dimension(name='space',coord=data.frame(lon=runif(50),lat=runif(50)),
      z=data.frame(elev=rnorm(50)))
space3D=dimension(name='space',coord=data.frame(x=runif(50),y=runif(50),z=runif(50)),
        z=data.frame(z=rnorm(50)))
# plot them
plot(time)
plot(space)
plot(space3D)
}
