% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{plotDAG}
\alias{plotDAG}
\title{Plot a DAG}
\usage{
plotDAG(
  DAG,
  fontSize = 12,
  opacity = 1,
  legend = TRUE,
  zoom = TRUE,
  arrows = TRUE,
  colourScale = NULL,
  ...
)
}
\arguments{
\item{DAG}{list. A DAG object, i.e. a list with 2 data frames (nodes and links)
typically resulting from a call of the function 'getDAG()'}

\item{fontSize}{numeric. Text size.}

\item{opacity}{numeric between 0 ad 1. Nodes opacity.}

\item{legend}{logical. Add legend?}

\item{zoom}{logical. Enable zoom?}

\item{arrows}{logical. Use arrows for links?}

\item{colourScale}{string. A JavaScript/D3 valid colorscale.}

\item{...}{other arguments passed to the function 'forceNetwork()'}
}
\value{
nothing - plotting function.
}
\description{
Plot the Directed Acyclic Graph  of a model
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=data.frame(Y=rlnorm(50)))
# Define parameters to be estimated: mu and sigma
param <- list(parameter(name='mu',init=0),parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu','sdlog=sigma'),par=param)
# get its DAG
dag<-getDAG(mod)
# plot it
plotDAG(dag)
}
