% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{plotMCMC.par}
\alias{plotMCMC.par}
\title{MCMC parameter plots}
\usage{
plotMCMC.par(mcmc, model, panelPerCol = 10)
}
\arguments{
\item{mcmc}{data frame, MCMC samples (typically from a call to readMCMC())}

\item{model}{Model, model object}

\item{panelPerCol}{Integer, max number of panels per column}
}
\description{
MCMC-based estimation of marginal posterior pdf for each parameter
}
