% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{plotMCMC.process}
\alias{plotMCMC.process}
\title{MCMC process plots}
\usage{
plotMCMC.process(mcmc, model)
}
\arguments{
\item{mcmc}{data frame, MCMC samples (typically from a call to readMCMC())}

\item{model}{Model, model object}
}
\description{
MCMC-based estimation of marginal posterior pdf for each process
}
