% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{plotMCMC.trace}
\alias{plotMCMC.trace}
\title{MCMC trace plots}
\usage{
plotMCMC.trace(mcmc, model = NULL, panelPerCol = 10)
}
\arguments{
\item{mcmc}{data frame, MCMC samples (typically from a call to readMCMC())}

\item{model}{Model, model object for smart coloring (black if NULL)}

\item{panelPerCol}{Integer, max number of panels per column}
}
\description{
Generate a trace of MCMC samples for all variables
}
