% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_process.R
\name{process}
\alias{process}
\title{Process constructor.}
\usage{
process(
  name,
  dim,
  dist = "Flat",
  par = NULL,
  iScalarPar = NULL,
  fMean = NULL,
  fCovar = NULL,
  init = NULL,
  constraint = 0
)
}
\arguments{
\item{name}{Character, name of the process.}

\item{dim}{Dimension object, on which the process is defined.}

\item{dist}{Character, name of the process (multivariate) distribution.}

\item{par}{Parameter, a list of parameter objects.}

\item{iScalarPar}{Integer vector, indices of scalar parameters in par.}

\item{fMean}{Character, formula to compute the mean vector.}

\item{fCovar}{Character, formula to compute the covariance matrix.}

\item{init}{Numeric or Character. If numeric, initial value of the process;
if Character, name of the dim-covariate used to initialise the process.}

\item{constraint}{Numeric, constraint index. 0 = none,
1 = process is centred, 2 = process is centred & scaled.}
}
\value{
An object of class 'process'.
}
\description{
Creates a new instance of a 'process' object
}
\examples{
# create a time dimension
time <- dimension(name='time',coord=data.frame(t=1:50))
# unconstrained time process
pro <- process(name='unconstrained',dim=time)
# IID Gaussian time process, flat priors on hyperparameters
param <- list(parameter(name='mean',init=0),
              parameter(name='sdev',init=1))
pro <- process(name='IID',dim=time,dist='Gaussian_IID',
               par=param,iScalarPar=c(1,2))
# AR(1) Gaussian time process, informative priors on hyperparameters
param<-list(parameter(name='cte',init=0,priorDist='Gaussian',priorPar=c(0,5)),
            parameter(name='sdev.innov',init=1,priorDist='FlatPrior+'),
            parameter(name='rho',init=0.5,priorDist='Uniform',priorPar=c(0,1)))
pro <- process(name='AR1',dim=time,dist='Gaussian_AR1',
               par=param,iScalarPar=c(1,2,3))
# Gaussian spatial process with zero mean
space <- dimension(name='space',coord=data.frame(x=runif(50),y=runif(50)))
param<-list(parameter(name='mean',init=0,priorDist='FIX'),
            parameter(name='range',init=5,priorDist='Exponential',priorPar=c(0,5)),
            parameter(name='sill',init=0.5,priorDist='FlatPrior+'))
pro <- process(name='spatialGP',dim=space,dist='Gaussian',
               par=param,fMean='mean',fCovar='(sill^2)*exp(-Dspace/range)')
# NNGP (Nearest-Neighbor Gaussian Process) with 5 neighbors max
space <- dimension(name='space',coord=data.frame(x=runif(50),y=runif(50)))
param<-list(parameter(name='mean',init=0,priorDist='FIX'),
            parameter(name='range',init=5,priorDist='Exponential',priorPar=c(0,5)),
            parameter(name='sill',init=0.5,priorDist='FlatPrior+'),
            parameter(name='maxNN',init=5,priorDist='FIX'))
pro <- process(name='NNGP',dim=space,dist='NNGP',par=param,iScalarPar=4,
               fMean='mean',fCovar='(sill^2)*exp(-Dspace/range)')
}
