% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{quickWriteDF}
\alias{quickWriteDF}
\title{Write a data frame to file)}
\usage{
quickWriteDF(DF, file)
}
\arguments{
\item{DF}{data frame to be written}

\item{file}{Character, file (full path)}
}
\value{
Nothing
}
\description{
Write a data frame to file)
}
\keyword{internal}
