% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{springFlow}
\alias{springFlow}
\title{Spring average flows in Eastern Australia}
\format{
An object of class \code{data.frame} with 64 rows and 43 columns.
}
\source{
\url{http://www.bom.gov.au/water/hrs/}
}
\usage{
springFlow
}
\description{
A data frame containing average flows during the
spring (September - October - November)
for 42 sites in Eastern Australia
over the period 1951-2014.
It has 64 rows (64 years from 1951 to 2014) and
43 variables (the year + average flow in m3/s at each of the 42 stations).
}
\keyword{datasets}
