% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUI.R
\name{STooDs}
\alias{STooDs}
\title{Run STooDs}
\usage{
STooDs(
  model,
  workspace = file.path(getwd(), "STooDs_workspace"),
  dir.exe = file.path(path.package("RSTooDs"), "bin"),
  name.exe = "STooDs",
  run = TRUE,
  mcmcOptions = mcmc()
)
}
\arguments{
\item{model}{Model object, the model to be run.}

\item{workspace}{Character, directory where config and result files are stored.}

\item{dir.exe}{Character, directory where STooDs.exe stands.}

\item{name.exe}{Character, name of the executable without extension ('STooDs' by default).}

\item{run}{Logical, run STooDs.exe? if FALSE, just write config files.}

\item{mcmcOptions}{MCMC object, MCMC options}
}
\value{
Nothing: just write config files and runs the executable.
}
\description{
Run STooDs.exe
}
