% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dataset.R
\name{toString.dataset}
\alias{toString.dataset}
\title{Dataset to string}
\usage{
\method{toString}{dataset}(x, ...)
}
\arguments{
\item{x}{Dataset object, dataset to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'dataset' into a ready-to-write vector of string
}
\examples{
toString(dataset(data.frame(Y=rlnorm(50))))
}
