% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dimension.R
\name{toString.dimension}
\alias{toString.dimension}
\title{Dimension to string}
\usage{
\method{toString}{dimension}(x, ...)
}
\arguments{
\item{x}{Dimension object, dimension to be converted}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'dimension' into a ready-to-write vector of string
}
\examples{
toString(dimension(name='time',coord=data.frame(t=1:50)))
}
