% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_distance.R
\name{toString.distance}
\alias{toString.distance}
\title{Distance to string}
\usage{
\method{toString}{distance}(x, ...)
}
\arguments{
\item{x}{Distance to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'distance' into a ready-to-write vector of string
}
\examples{
toString(distance(funk='Euclidean'))
}
