% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_mcmc.R
\name{toString.mcmc}
\alias{toString.mcmc}
\title{mcmc to string}
\usage{
\method{toString}{mcmc}(x, ...)
}
\arguments{
\item{x}{mcmc object, object to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'mcmc' into a ready-to-write vector of string
}
\examples{
toString(mcmc())
}
