% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{toString.model}
\alias{toString.model}
\title{Model to string}
\usage{
\method{toString}{model}(x, ...)
}
\arguments{
\item{x}{Model to be converted.}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'model' into a ready-to-write vector of string
}
\examples{
# Generate a fake dataset
dat <- dataset(Y=data.frame(Y=rlnorm(50)))
# Define parameters to be estimated: mu and sigma
param <- list(parameter(name='mu',init=0),parameter(name='sigma',init=1))
# Define model
mod <- model(dataset=dat,parentDist='LogNormal',
             formula=c('meanlog=mu','sdlog=sigma'),par=param)
toString(mod)
}
