% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_process.R
\name{toString.process}
\alias{toString.process}
\title{Process to string}
\usage{
\method{toString}{process}(x, ...)
}
\arguments{
\item{x}{process to be converted}

\item{...}{Optional arguments.}
}
\value{
A string ready to be printed or written.
}
\description{
Convert an object of class 'process' into a ready-to-write vector of string
}
\examples{
time <- dimension(name='time',coord=data.frame(t=1:50))
toString(process('unconstrained',time))
}
